EuclideanDistance now returns int instead of double

+++ nightlight/nightlight/AiUtil.h
@@ -37,7 +37,7 @@ static int ManhattanDistance(Node n1, Node n2)
 
 static int EuclideanDistance(Node n1, Node n2)
 {
-	return sqrt((pow((n1.x - n2.x), 2) + pow((n1.y - n2.y), 2)));
+	return (int)sqrt((pow((n1.x - n2.x), 2) + pow((n1.y - n2.y), 2)));
 }
 
 static int ChebyshevDistance(Node n1, Node n2)
